requirements.txt
django
psycopg2
psycopg2-binary
pillow

django-phonenumber-field